Move bulk handling of parameter info to /parameters/bulk
